Full App Refactor

+++ src/pages/settings/Settings.tsx
@@ -1,19 +1,24 @@
-import { Save } from '@mui/icons-material';
+import { Delete, Edit, Save } from '@mui/icons-material';
 import { LoadingButton } from '@mui/lab';
-import { Autocomplete, Button, TextField } from '@mui/material';
+import { Autocomplete, Button, IconButton, TextField } from '@mui/material';
 import { useFormik } from 'formik';
-import $ from 'jquery';
-import React, { lazy, useEffect, useState } from 'react';
+import React, { lazy, useState } from 'react';
 import { Modal } from 'react-bootstrap';
 import Swal from 'sweetalert2';
 import withReactContent from 'sweetalert2-react-content';
 import * as yup from 'yup';
 import { settingOptions } from '../../constants';
-import { useDeleteSettingMutation, useSettingsQuery, useUpsertSettingMutation } from '../../features/notifications/notificationsAPI';
-import Master from '../../layouts/Master';
-import { SettingType } from '../../utils/types';
+import {
+    useDeleteSettingMutation,
+    useSettingsQuery,
+    useUpsertSettingMutation
+} from '../../features/notifications/notificationsAPI';
+import { SettingType } from 'helpers/types';
+import { SectionLoader } from '../../components/Loader';
+import { toast } from '../../helpers/utils';
+import { SectionError } from '../../components/Error';
 
-const SettingsTable = lazy(() => import('./SettingsTable'));
+const DataTable = lazy(() => import('components/common/datatable'));
 
 const MySwal = withReactContent(Swal);
 
@@ -29,44 +34,32 @@ const validationSchema = yup.object({
 });
 
 const Settings = () => {
-    const [settings, setSettings] = useState<SettingType[]>(null!);
     const [settingValues, setSettingValues] = useState<string[]>([]);
     const [showModal, setShowModal] = useState(false);
 
-    let { data, error } = useSettingsQuery();
-    const [upsertSetting, result] = useUpsertSettingMutation()
-    const [deleteSetting] = useDeleteSettingMutation()
-
-    useEffect(() => {
-        if (data) {
-            setSettings(data);
-
-            setTimeout(() => {
-                $('#table_id').DataTable({
-                    retrieve: true,
-                    columnDefs: [
-                        { orderable: false, targets: [0, 5] }
-                    ],
-                    lengthMenu: [[7, 10, -1], [7, 10, "All"]]
-                });
-            }, 500);
-        }
-    }, [data]);
+    let {data: settings, error, isLoading, isSuccess} = useSettingsQuery();
+    const [upsertSetting, result] = useUpsertSettingMutation();
+    const [deleteSetting] = useDeleteSettingMutation();
 
     const formik = useFormik({
-        initialValues: { type: "", value: "" },
+        initialValues: {type: "", value: ""},
         validationSchema: validationSchema,
         onSubmit: async values => {
-            await upsertSetting(values)
-            
-            if(result.isSuccess) {
-                setShowModal(false);
-            }
+            const setting = await upsertSetting(values).unwrap();
+
+            setShowModal(false);
+
+            if(setting?.id) toast({
+                msg: `Setting ${formik.dirty ? "Updated" : "Created"}!`,
+                type: "success"
+            });
+
+            if (result.error) toast({msg: result.error.toString(), type: 'warning'});
         },
     });
 
     const handleCreate = () => {
-        formik.resetForm()
+        formik.resetForm();
 
         setShowModal(true);
     };
@@ -90,28 +83,64 @@ const Settings = () => {
                 confirmButtonText: 'Yes, delete it!',
                 showLoaderOnConfirm: true
             }).then(async result => {
-                if (result.isConfirmed) {
-                    await deleteSetting(String(setting.id))
-
-                    // if (loading) MySwal.showLoading();
-                }
+                if (result.isConfirmed) await deleteSetting(String(setting.id));
             });
         }
     };
 
     return (
-        <Master error={error}>
-            <div className="row g-3 mb-3">
-                <div className="col-xxl-12 col-md-12">
-                    <SettingsTable settings={settings} onCreateRow={handleCreate} onDeleteRow={handleDelete}
-                        onUpdateRow={handleUpdate} setSettings={setSettings} />
-                </div>
-            </div>
+        <>
+            {error && <SectionError error={error}/>}
+            {
+                !isLoading && isSuccess && settings
+                    ? <div className="row g-3 mb-3">
+                        <div className="col-xxl-12 col-md-12">
+                            <DataTable tableClassName={'table-sm'} perPage={5} columns={[
+                                {
+                                    accessor: 'type',
+                                    Header: 'Name',
+                                    Cell: (rowData: any) => {
+                                        const {type} = rowData.row.original;
+
+                                        return <span>{(type.replaceAll('_', ' ')).toUpperCase()}</span>;
+                                    }
+                                },
+                                {
+                                    accessor: 'value',
+                                    Header: 'Value'
+                                },
+                                {
+                                    accessor: 'actions',
+                                    disableSortBy: true,
+                                    className: 'text-end',
+                                    Cell: (rowData: any) => {
+                                        return (
+                                            <div className={'text-end'}>
+                                                <IconButton onClick={() => handleUpdate(rowData.row.original)}
+                                                            size={"small"}
+                                                            color={"primary"}>
+                                                    <Edit fontSize={'small'}/>
+                                                </IconButton>
+                                                <IconButton onClick={() => handleDelete(rowData.row.original)}
+                                                            size={"small"}
+                                                            color={"error"}>
+                                                    <Delete fontSize={'small'}/>
+                                                </IconButton>
+                                            </div>
+                                        );
+                                    }
+                                }
+                            ]} data={settings.map(setting => setting)} title={'Settings'} bulkActions
+                                       onCreateRow={handleCreate}/>
+                        </div>
+                    </div>
+                    : <SectionLoader/>
+            }
 
             <Modal show={showModal} onHide={() => setShowModal(false)} contentClassName={'position-relative'}>
                 <div className="position-absolute top-0 end-0 mt-2 me-2 z-index-1">
                     <button className="btn-close btn btn-sm btn-circle d-flex flex-center transition-base"
-                        data-bs-dismiss="modal" aria-label="Close" onClick={() => setShowModal(false)} />
+                            data-bs-dismiss="modal" aria-label="Close" onClick={() => setShowModal(false)}/>
                 </div>
                 <Modal.Body className={'modal-body p-0'}>
                     <div className="rounded-top-lg py-3 ps-4 pe-6 bg-light">
@@ -125,32 +154,32 @@ const Settings = () => {
                             {result.error}
                             <div className="mb-3">
                                 <Autocomplete options={settingOptions.map(opt => opt.type)} freeSolo
-                                    value={formik.values.type}
-                                    onChange={(event, newValue) => {
-                                        const settingValues = getSettingValuesByName(String(newValue));
-                                        setSettingValues(settingValues);
-                                        formik.setFieldValue("type", newValue, true);
-                                    }}
-                                    renderInput={(params) => (
-                                        <TextField{...params} size={"small"} label="Type"
-                                            placeholder="Setting type..." value={formik.values.type}
-                                            error={formik.touched.type && Boolean(formik.errors.type)}
-                                            helperText={formik.touched.type && formik.errors.type} />
-                                    )}
+                                              value={formik.values.type}
+                                              onChange={(event, newValue) => {
+                                                  const settingValues = getSettingValuesByName(String(newValue));
+                                                  setSettingValues(settingValues);
+                                                  formik.setFieldValue("type", newValue, true);
+                                              }}
+                                              renderInput={(params) => (
+                                                  <TextField{...params} size={"small"} label="Type"
+                                                            placeholder="Setting type..." value={formik.values.type}
+                                                            error={formik.touched.type && Boolean(formik.errors.type)}
+                                                            helperText={formik.touched.type && formik.errors.type}/>
+                                              )}
                                 />
                             </div>
                             <div className="mb-3">
                                 <Autocomplete options={settingValues} freeSolo value={formik.values.value}
-                                    onChange={(event, newValue) => {
-                                        formik.setFieldValue("value", newValue, true);
-                                    }}
-                                    renderInput={(params) => (
-                                        <TextField{...params} size={"small"} label="Value"
-                                            placeholder="Setting value..."
-                                            value={formik.values.value}
-                                            error={formik.touched.value && Boolean(formik.errors.value)}
-                                            helperText={formik.touched.value && formik.errors.value} />
-                                    )}
+                                              onChange={(event, newValue) => {
+                                                  formik.setFieldValue("value", newValue, true);
+                                              }}
+                                              renderInput={(params) => (
+                                                  <TextField{...params} size={"small"} label="Value"
+                                                            placeholder="Setting value..."
+                                                            value={formik.values.value}
+                                                            error={formik.touched.value && Boolean(formik.errors.value)}
+                                                            helperText={formik.touched.value && formik.errors.value}/>
+                                              )}
                                 />
                             </div>
                         </form>
@@ -158,15 +187,15 @@ const Settings = () => {
                 </Modal.Body>
                 <Modal.Footer>
                     <Button size={'small'} variant={'outlined'} onClick={() => setShowModal(false)} color={'inherit'}
-                        data-bs-dismiss="modal">Cancel</Button>
+                            data-bs-dismiss="modal">Cancel</Button>
                     <LoadingButton size="small" color="primary" loading={result.isLoading} loadingPosition="end"
-                        onClick={() => formik.submitForm()}
-                        endIcon={<Save />} variant="contained">
+                                   onClick={() => formik.submitForm()}
+                                   endIcon={<Save/>} variant="contained">
                         {formik.dirty ? "Update" : "Create"}
                     </LoadingButton>
                 </Modal.Footer>
             </Modal>
-        </Master>
+        </>
     );
 };
